generate_custom_delay
