+++ Keil/generate_custom_delay/config.txt
@@ -3,4 +3,4 @@ Project config information:
 1. Manage runtime environment: Select CMSIS:Core and Device:Startup
 2. Options for Target <...>: 
 	a. Choose Stellaris ICDI under Use in Debug tab.
-	b. C/C++: choose C11 (language C), MISRA compatible (Warnings), add inc folder of project.
+	b. C/C++: choose C11 (language C), MISRA compatible (Warnings), add inc folder of project, define SECONDS/MILLISECONDS as per need.
