+++ Keil/generate_custom_delay/readme.txt
@@ -1,0 +1,3 @@
+Generates maximum delay of 1.00352 seconds, minimum delay of 4.096 milliseconds when multiplier(m) = 1.
+Define either SECONDS or MILLISECONDS to generate a suitable delay. Adjust seconds/milliseconds value in header file.
+Maximum checked delay = 5 minutes (300 seconds).
