--- Keil/generate_delay/main.c
@@ -1,11 +1,0 @@
-#include "generatedelay.h"
-
-int main(){
-	use_indicator();
-	//set_miliseconds(100);
-	set_seconds(30);
-	disable_timer();
-	while(1){
-		//TODO
-	}
-}
