--- Keil/generate_delay/readme.txt
@@ -1,2 +1,0 @@
-Generates maximum delay of 1.00352 seconds, minimum delay of 4.096 miliseconds when multiplier(m) = 1.
-Maximum checked delay = 5 minutes (300 seconds).
